Fix LZMA file decoding/encoding
LZMA typically contains the uncompressed size of the file in the header
(with each byte set to -1 if the size is unknown), but this is stripped
from the LZMA-compressed files in the game cache, as the uncompressed
size is stored in the container header already. This changeset
accommodates that behaviour.